backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pdfplumber>=0.11.0
google-genai>=0.1.0
edge-tts>=6.1.12
boto3>=1.34.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pydub>=0.25.1
